requirements.txt
openai
python-dotenv
gradio
transformers
torch
gtts
